perf: melhorias implementadas na aplicacao
